requirements.txt
bs4
pyfiglet
python-aiml
youtube-dl
redis
requests
telepot